move bifrost_current up and dont have a dedicated hdf5 directory
