+++ crates/plotinator-test-util/src/lib.rs
@@ -7,6 +7,6 @@ pub mod test_file_defs;
 pub use {
     std::fs,
     std::io,
-    test_file_defs::{hdf5::*, legacy_generator::*, mbed_motor_control::*, wasp200::*},
+    test_file_defs::{bifrost_current::*, legacy_generator::*, mbed_motor_control::*, wasp200::*},
     testresult::TestResult,
 };
