+++ crates/plotinator-test-util/src/test_file_defs.rs
@@ -48,8 +48,8 @@ macro_rules! define_utf8_test_file {
     };
 }
 
+pub mod bifrost_current;
 pub mod frame_altimeters;
-pub mod hdf5;
 pub mod legacy_generator;
 pub mod mbed_motor_control;
 pub mod wasp200;
