--- crates/plotinator-test-util/src/test_file_defs/hdf5.rs
@@ -1,1 +1,0 @@
-pub mod bifrost_current;
